Moved different classes into different files
