+++ lib/model/budget_control.dart
@@ -12,7 +12,7 @@ import 'package:budgetflow/model/history/history.dart';
 import 'package:budgetflow/model/history/month_time.dart';
 import 'package:flutter/material.dart';
 
-import 'budget/category/budget_category.dart';
+import 'budget/category/category.dart';
 import 'history/month.dart';
 
 class BudgetControl implements Control {
